Womensafety-main/Womensafety-main/requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
werkzeug==3.0.1
twilio==8.14.0 